implementation with unreachable answers removed.
