+++ src/main/app/controller/PostController.ts
@@ -67,7 +67,7 @@ export class PostController<T extends AnyObject> {
       }
     }
 
-    const nextUrl = req.session.errors.length > 0 ? req.url : this.getNextStep(req, req.session.userCase);
+    const nextUrl = req.session.errors.length > 0 ? req.url : getNextStepUrl(req, req.session.userCase);
 
     req.session.save(err => {
       if (err) {
@@ -75,10 +75,6 @@ export class PostController<T extends AnyObject> {
       }
       res.redirect(nextUrl);
     });
-  }
-
-  protected getNextStep(req: AppRequest<T>, formData: Partial<CaseWithId>): string {
-    return getNextStepUrl(req, formData);
   }
 
   protected async save(req: AppRequest<T>, formData: Partial<Case>, eventName: string): Promise<CaseWithId> {
